platform/examples/langchain/requirements.txt
langchain>=0.3.0
langchain-core>=1.2.0,<2.0.0
langchain-classic>=0.3.0,<1.0.0

langchain-openai>=1.1.0,<2.0.0
langchain-anthropic>=1.3.0,<2.0.0
langchain-google-genai>=4.1.0,<5.0.0

requests>=2.32.0,<3.0.0

python-dotenv>=1.0.0,<2.0.0
